% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_inicial.R
\name{carrega_latitude}
\alias{carrega_latitude}
\title{Carrega o vetor de latitudes}
\usage{
carrega_latitude(modelo, nome_latitude = "latitude")
}
\arguments{
\item{modelo}{O Netcdf já carregado com a funcao (carrega_modelo)}

\item{nome_latitude}{latitude, lat, latitude}
}
\value{
o vetor de latitude carregado
}
\description{
Carrega o vetor de latitudes
}
\examples{
latitudes <- carrega_latitude(gfs,nome_latitude="lat")
          latitudes <- carrega_latitude(gfs)
}
