% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_inicial.R
\name{carrega_longitude}
\alias{carrega_longitude}
\title{Load the longitude vector}
\usage{
carrega_longitude(modelo, nome_longitude = "longitude")
}
\arguments{
\item{modelo}{The loaded netcdfile file (using the function carrega_modelo)}

\item{nome_longitude}{The longitude name as it is in the Netcdf file (e.g. longitude, lon, LONGITUDE)}
}
\value{
Returns the longitude vector
}
\description{
Converts from 0-360 to -180 to 180 degrees
}
\examples{
longitudes <- carrega_longitude(gfs,nome_longitude="lon")
          longitudes <- carrega_longitude(gfs)
}
