% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_inicial.R
\name{carrega_modelo}
\alias{carrega_modelo}
\title{Load the netcdf file}
\usage{
carrega_modelo(arquivo)
}
\arguments{
\item{arquivo}{Nome (e caminho) do arquivo em netcdf}
}
\value{
The loaded netecdf file
}
\description{
Abre o arquivo com a funcao nc_open do ncdf4
}
\examples{
arq <- netcdfplot::gfs_example_file()
modelo <- netcdfplot::carrega_modelo(arq)
}
