% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes_gis.R
\name{carrega_shape}
\alias{carrega_shape}
\title{carrega_shape}
\usage{
carrega_shape(mapa = NA)
}
\arguments{
\item{mapa}{Entrar com o nome e local do shape ou "Brasil", "America_do_Sul", "Mundo", "Pianco", "PPA", "Piracicaba"}
}
\value{
shape
}
\description{
carrega_shape
}
