% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dimensoes.R
\name{carrega_tempo}
\alias{carrega_tempo}
\title{carrega_tempo}
\usage{
carrega_tempo(nc, time = "time")
}
\arguments{
\item{nc}{Arquivo carregado em netcdf}

\item{time}{Entrar com time ou valid_time, dependendo como vem no nc}
}
\value{
Esta função carrega o tempo de um arquivo Netcdf no formato adequado
}
\description{
Entrar com o netcdf já carregado
}
\examples{
carrega_tempo(gfs)
          carrega_tempo(era5,time="valid_time")
}
