% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_inicial.R
\name{carrega_variavel}
\alias{carrega_variavel}
\title{Load the variable data (e.g. temperature, precipitation) of a netcdf file}
\usage{
carrega_variavel(modelo, nome_variavel)
}
\arguments{
\item{modelo}{The loaded netcdfile file (using the function carrega_modelo)}

\item{nome_variavel}{Name of the variable (use shownames() function to find it)}
}
\value{
THe variable array
}
\description{
Load the variable data (e.g. temperature, precipitation) of a netcdf file
}
\examples{
temperatura <- carrega_variavel(gfs,"tmp2m")
}
