% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dimensoes.R
\name{explainname}
\alias{explainname}
\title{explainname}
\usage{
explainname(nc, var)
}
\arguments{
\item{nc}{Arquivo carregado em netcdf}

\item{var}{Nome da variável}
}
\value{
Detalhes sobre a variável
}
\description{
Fornece detalhes sobre a variável do Netcdf
}
\examples{
explainname(gfs,"precipitation")
          explainname(gfs,"tmp2m")
}
