% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes_gis.R
\name{extractfromshape}
\alias{extractfromshape}
\title{extractfromshape}
\usage{
extractfromshape(raster, shape, fun = "mean", quantil = NA)
}
\arguments{
\item{raster}{Raster cuja média ou outra função estatística vai ser extraída}

\item{shape}{Dentro do shape que vai extrair a média do raster}

\item{fun}{A função estatística mean, min ou max}

\item{quantil}{Valor entre 0 e 1 do quantil para obter dentro da área do shape}
}
\value{
Média ou função estatística ou percentil dentro da área do shape
}
\description{
Extrai a media ou outro de um raster dentro da area de um poligono
}
