% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_example.R
\name{gfs_example_file}
\alias{gfs_example_file}
\title{Caminho para o NetCDF de exemplo}
\usage{
gfs_example_file()
}
\value{
Caminho completo para o arquivo .nc
}
\description{
Caminho para o NetCDF de exemplo
}
