% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dimensoes.R
\name{lat2y}
\alias{lat2y}
\title{lat2y}
\usage{
lat2y(vetor_lat, lati)
}
\arguments{
\item{vetor_lat}{Vetor com as latitude}

\item{lati}{Latitude a set convertida para y}
}
\value{
posição do vetor com a latitude mais próxima da solicitada
}
\description{
Converte a latitude em coordenada y
}
\examples{
lat2y(latitudes,-23.5)
}
