% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes_gis.R
\name{matrix2raster}
\alias{matrix2raster}
\title{Converts a matrix to raster in WGS85 longlat}
\usage{
matrix2raster(matrix, vlat, vlon)
}
\arguments{
\item{matrix}{Matriz de dimensão vlon X vlat}

\item{vlat}{Vetor de latitudes}

\item{vlon}{Vetor de longitudes}
}
\value{
raster
}
\description{
Converts a matrix to raster in WGS85 longlat
}
\examples{
matrix2raster(matriz,vlat,vlon)
}
