% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes_gis.R
\name{media_area}
\alias{media_area}
\title{Extracts the mean of a matrix within a shapefile}
\usage{
media_area(matrix, mapa, vlat, vlon, fun = "mean", quantil = NA)
}
\arguments{
\item{matrix}{Matrix of dimension vlon X vlat}

\item{mapa}{Definido para Brasil, America_do_Sul, Mundo, Pianco, PPA, Piracicaba}

\item{vlat}{Latitude vector}

\item{vlon}{Longitudes vector}

\item{fun}{A função estatística mean, min ou max}

\item{quantil}{Valor entre 0 e 1 do quantil para obter dentro da área do shape}
}
\value{
The mean within a shapefile
}
\description{
Extracts the mean of a matrix within a shapefile
}
\examples{
\dontrun{
media_area(temperatura_matriz, mapa = "Brasil", vlat = latitudes, vlon = longitudes, fun="max")
media_area(temperatura_matriz, mapa = "/storage/dados/shapes/shape_bacia_SF/bacia_sao_francisco.shp", vlat = latitudes, vlon = longitudes, quantil=0.9)
}
}
