% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdfplot-package.R
\docType{package}
\name{netcdfplot}
\alias{netcdfplot}
\title{netcdfplot}
\description{
Ferramentas para plotar e analisar dados em Netcdf
}
\details{
Sandra I. Saad
sandraisaad@gmail.com
}
\section{Exemplos}{

gfs <- carrega_modelo("/data/gfs_2019092500mod.nc")
longitudes <- carrega_longitude(gfs)
latitudes <- carrega_latitude(gfs)
temperatura <- carrega_variavel(gfs,"tmp2m")
temperatura_tempo1 <-temperatura[,,1]
plota_mapa(temperatura_tempo1, vlon = longitudes, vlat = latitudes, mapa_brasil= T, mapa_amsul = T, paleta = "YlGnBu")
media_area(temperatura_tempo1, mapa = "Brasil",vlat = latitudes, vlon = longitudes,fun = "max")
}

