% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plota.R
\name{plota_mapa}
\alias{plota_mapa}
\title{Plota um mapa}
\usage{
plota_mapa(
  array_var2d,
  titulo = "",
  legenda = "",
  vlon,
  vlat,
  loni = NA,
  lonf = NA,
  lati = NA,
  latf = NA,
  mapa = NA,
  paleta = "RdBu",
  paleta_rev = F,
  mapa_brasil = F,
  mapa_amsul = F,
  mapa_mundo = F,
  intervalos = NA,
  interpolate = T,
  legenda_intervalos = F,
  contorno = F
)
}
\arguments{
\item{array_var2d}{Matriz (em 2 dimensões) com o campo a ser plotado}

\item{titulo}{Título do mapa (opcional)}

\item{legenda}{Nome em cima da legenda (opcional)}

\item{vlon}{Vetor com as longitudes correspondente a array_var2d}

\item{vlat}{Vetor com as latitudes}

\item{loni}{Longitude inicial para o corte do mapa (opcional)}

\item{lonf}{final}

\item{lati}{Latitude  inicial}

\item{latf}{Latitude  final}

\item{mapa}{Mapa a ser mostrado}

\item{paleta}{Paleta de cores a ser utilizada. Default: "RdBu".Ver outras em http://www.sthda.com/english/wiki/colors-in-r}

\item{paleta_rev}{Se T, inverte a paleta de cores}

\item{mapa_brasil}{Se T, acrescenta o mapa do Brasil e os estados}

\item{mapa_amsul}{Se T, acrescenta o mapa da América do Sul}

\item{mapa_mundo}{Se T, acrescenta o mapa dos países do mundo}

\item{intervalos}{Um vetor com os intervalo de dados a ser considerado. Vale se legenda_intervalos=T}

\item{interpolate}{Se F, mostra pixelado e não interpolado}

\item{legenda_intervalos}{Se T, mostra a legenda em intervalos}

\item{contorno}{Se T, mostra os contornos do plot}
}
\description{
Plota um mapa
}
\examples{

plota_mapa(prec,titulo = expression(paste("Precipitation (mm yr"^"-1",")")),legenda="Precipitação (mm)", vlon=longitudes,vlat=latitudes,loni=-39,lonf=-36,lati=-7.78,latf=-5, mapa = "PPA", mapa_brasil = T,mapa_amsul = T, paleta_rev = F, paleta = "YlGnBu", legenda_intervalos = F, ,contorno = F)

plota_mapa(temperatura,vlon=longitudes,vlat=latitudes)

}
