% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plota.R
\name{plota_mapa_vento}
\alias{plota_mapa_vento}
\title{Plots wind vectorial field}
\usage{
plota_mapa_vento(
  array_u2d,
  array_v2d,
  titulo = "",
  legenda = "",
  vlon,
  vlat,
  loni = NA,
  lonf = NA,
  lati = NA,
  latf = NA,
  mapa = NA,
  mapa_brasil = F,
  mapa_amsul = F,
  mapa_mundo = F,
  intervalos = NA,
  legenda_intervalos = F,
  skip = 3
)
}
\arguments{
\item{array_u2d}{Matriz da componente zonal do vento, u, em 2 dimensões}

\item{array_v2d}{Matriz da componente meridional do vento, v, em 2 dimensões}

\item{titulo}{Título do mapa (opcional)}

\item{legenda}{Nome em cima da legenda (opcional)}

\item{vlon}{Vetor com as longitudes correspondente a array_var2d}

\item{vlat}{Vetor com as latitudes}

\item{loni}{Longitude inicial para o corte do mapa (opcional)}

\item{lonf}{final}

\item{lati}{Latitude  inicial}

\item{latf}{Latitude  final}

\item{mapa}{Mapa a ser mostrado}

\item{mapa_brasil}{Se T, acrescenta o mapa do Brasil e os estados}

\item{mapa_amsul}{Se T, acrescenta o mapa da América do Sul}

\item{mapa_mundo}{Se T, acrescenta o mapa dos países do mundo}

\item{intervalos}{Um vetor com os intervalo de dados a ser considerado. Vale se legenda_intervalos=T}

\item{legenda_intervalos}{Se T, mostra a legenda em intervalos}

\item{skip}{Pula alguns pontos de grades para nao ficar poluido. Padrao 3}
}
\description{
Plots wind vectorial field
}
\examples{

plota_mapa_vento(u10m_t1,v10m_t1,vlon = longitudes, vlat = latitudes)

}
