% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plota.R
\name{salva_figura}
\alias{salva_figura}
\title{Save image}
\usage{
salva_figura(plt, file_name, width = 2500, height = 2000, salva_svg = F)
}
\arguments{
\item{plt}{A figura plotada no R}

\item{file_name}{Nome da Figura sem extensão}

\item{width}{Largura a ser salva}

\item{height}{Altura a ser salva}

\item{salva_svg}{Se T salva em svg também}
}
\description{
Plot image and save in jpg and svg (if chosen)
}
\examples{
\dontrun{
salva_figura(plt,"figura_temperatura",width=2500,height=2000,salva_svg=F)
}
}
