% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dimensoes.R
\name{shownames}
\alias{shownames}
\title{shownames}
\usage{
shownames(nc)
}
\arguments{
\item{nc}{Arquivo carregado em netcdf}
}
\value{
nome das variaveis do arquivo
}
\description{
shownames
}
\examples{
shownames(gfs)
}
