% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_inicial.R
\name{carrega_latitude}
\alias{carrega_latitude}
\title{Load the latitude vector}
\usage{
carrega_latitude(modelo, nome_latitude = "latitude")
}
\arguments{
\item{modelo}{The loaded netcdfile file (using the function carrega_modelo)}

\item{nome_latitude}{The latitude name as it is in the Netcdf file (e.g. latitude, lat, LATITUDE)}
}
\value{
Returns the latitude vector
}
\description{
Load the latitude vector
}
\examples{
latitudes <- carrega_latitude(gfs,nome_latitude="lat")
          latitudes <- carrega_latitude(gfs)
}
