% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_inicial.R
\name{carrega_longitude}
\alias{carrega_longitude}
\title{Carrega o vetor de longitudes}
\usage{
carrega_longitude(modelo, nome_longitude = "longitude")
}
\arguments{
\item{modelo}{O Netcdf já carregado com a funcao (carrega_modelo)}

\item{nome_longitude}{longitude, lon, LONGITUDE}
}
\value{
o vetor de longitude carregado
}
\description{
Converte 0 a 360 em -180 a 180
}
\examples{
longitudes <- carrega_longitude(gfs,nome_longitude="lon")
          longitudes <- carrega_longitude(gfs)
}
