% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_inicial.R
\name{carrega_modelo}
\alias{carrega_modelo}
\title{carrega o modelo (o arquivo em Netcdf)}
\usage{
carrega_modelo(arquivo)
}
\arguments{
\item{arquivo}{Nome (e caminho) do arquivo em netcdf}
}
\value{
O modelo carregado
}
\description{
Abre o arquivo com a funcao nc_open do ncdf4
}
\examples{
gfs <- carrega_modelo("/data/gfs_2019092500mod.nc")
}
