% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dimensoes.R
\name{carrega_tempo}
\alias{carrega_tempo}
\title{Load date/time vector}
\usage{
carrega_tempo(nc, time = "time")
}
\arguments{
\item{nc}{The loaded netcdfile file (using the function carrega_modelo)}

\item{time}{Name of the time variable in the NetCDF file. Use
\code{"time"} (default) or \code{"valid_time"}.}
}
\value{
A date/time vector in POSIXct format
}
\description{
Returns the date/time vector from a NetCDF file in POSIXct format
}
\examples{
\dontrun{
carrega_tempo(gfs)
carrega_tempo(era5,time="valid_time")
}
}
