% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_inicial.R
\name{carrega_variavel}
\alias{carrega_variavel}
\title{Carrega o dado da variavel de um netcdf}
\usage{
carrega_variavel(modelo, nome_variavel)
}
\arguments{
\item{modelo}{O Netcdf já carregado com a funcao (carrega_modelo)}

\item{nome_variavel}{Nome da variavel}
}
\value{
O array da variavel (temperatura, precipitacao, etc)
}
\description{
Carrega o dado da variavel de um netcdf
}
\examples{
temperatura <- carrega_variavel(gfs,"tmp2m")
}
