% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dimensoes.R
\name{explainname}
\alias{explainname}
\title{Provides details about the requested variable in the netcdf file}
\usage{
explainname(nc, var)
}
\arguments{
\item{nc}{The loaded netcdfile file (using the function carrega_modelo)}

\item{var}{Variable name}
}
\value{
Details about the requested variable
}
\description{
Provides details about units and others
}
\examples{
\dontrun{
explainname(gfs,"precipitation")
          explainname(gfs,"tmp2m")
}
}
