% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_example.R
\name{gfs_example_file}
\alias{gfs_example_file}
\title{Find the NetCDF file provided for testing}
\usage{
gfs_example_file()
}
\value{
The full path to the example .nc file.
}
\description{
Find the NetCDF file provided for testing
}
\examples{
arq <- gfs_example_file()
}
