% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dimensoes.R
\name{lat2y}
\alias{lat2y}
\title{Converts a given Latitude to an y index}
\usage{
lat2y(vetor_lat, lati)
}
\arguments{
\item{vetor_lat}{Latitude vector}

\item{lati}{Latitude value to be converted to an y index}
}
\value{
The index of the latitude in \code{vetor_lat} that is
closest to the requested value.
}
\description{
Converts a latitude value to its corresponding y index (from 1 to the length of the latitude vector).
}
\examples{
\dontrun{
lat2y(latitudes,-23.5)
}
}
