% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dimensoes.R
\name{lon2x}
\alias{lon2x}
\title{lon2x}
\usage{
lon2x(vetor_lon, loni)
}
\arguments{
\item{vetor_lon}{Vetor com as longitudes}

\item{loni}{Longitude a set convertida para x}
}
\value{
posição do vetor com a longitude mais próxima da solicitada
}
\description{
Converte a longitude em coordenada x
}
\examples{
lon2x(longitudes,-46.5)
}
