% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes_gis.R
\name{media_area}
\alias{media_area}
\title{media_area}
\usage{
media_area(matrix, mapa, vlat, vlon, fun = "mean", quantil = NA)
}
\arguments{
\item{matrix}{Matriz de dimensão vlon X vlat}

\item{mapa}{Definido para Brasil, America_do_Sul, Mundo, Pianco, PPA, Piracicaba}

\item{vlat}{Vetor de latitudes}

\item{vlon}{Vetor de longitudes}

\item{fun}{A função estatística mean, min ou max}

\item{quantil}{Valor entre 0 e 1 do quantil para obter dentro da área do shape}
}
\value{
A média dentro do shape
}
\description{
Extrai a media de uma matriz em um shape. Converte p/ raster primeiro
}
\examples{
media_area(temperatura_matriz, mapa = "Brasil", vlat = latitudes, vlon = longitudes, fun="max")
media_area(temperatura_matriz, mapa = "/storage/dados/shapes/shape_bacia_SF/bacia_sao_francisco.shp", vlat = latitudes, vlon = longitudes, quantil=0.9)
}
