% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dimensoes.R
\name{shownames}
\alias{shownames}
\title{Shows the names of the variables available in the Netcdf file}
\usage{
shownames(nc)
}
\arguments{
\item{nc}{The loaded netcdfile file (using the function carrega_modelo)}
}
\value{
the names of the variables available in the Netcdf file
}
\description{
Shows the names of the variables available in the Netcdf file
}
\examples{
shownames(gfs)
}
